Moved unused components to incomplete-components
