+++ frontend/src/components/incomplete-components/main-view-modals/manager-room.js
