+++ frontend/src/components/incomplete-components/main-view-modals/stats.js
